utf8-extended: fix endian-less + fix utf-8 format checking
The fread() and fwrite() calls now reads and writes by n-bytes blocks
instead of byte-by-byte.

The fclose() function is now called on failure exits.

The utf-8 format checking on decoding had a bug because of the use of
the xor operator. That would have not signaled an error for utf-8 bytes
starting with the 2-bit string "00". This is now fixed.