interrato.dev: add support for protected notes
Protected notes are encrypted using age (https://age-encryption.org/),
each with its own passphrase. The corresponding HTML file is excluded
from version control (see the .gitignore file in the notes directory).
The "protect-note" recipe in the justfile generates a proxy HTML file
to let Tailwind know which CSS classes to include, and then calls age
to encrypt the specified note.

Metadata about protected notes (title, date, etc.) and their static
assets are not encrypted. The purpose of protected notes is to prevent
the contents of various notes that are intended for strictly personal
use or for a small group of people from being fully public online.